requirements.txt
matplotlib>=3.3.4
numpy>=1.20.1
opencv-python==4.5.1.48
Pillow==7.2.0
pre-commit==2.10.0
pytest==6.2.2
selenium==3.141.0
tensorflow-gpu==2.4.0
webdriver-manager==3.3.0
